feat: resource select drawer shared ui
